backend/requirements.txt
# RTI Voice Assistant - Backend Requirements
# Fully functional prototype with LLM services

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS SDK (for DynamoDB and S3 only)
boto3==1.34.0
botocore==1.34.0

# LLM Providers
groq==0.4.1
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# PDF Generation
reportlab==4.0.7

# Audio processing
pydub==0.25.1

# For handling audio files
python-multipart==0.0.6

# gTTS for Kannada support
gTTS==2.5.0
